Add md5 check on htaccess recommendation

+++ include/classes.php
@@ -772,7 +772,7 @@ class mf_cache
 	{
 		$folder = $data['path']."/".$data['child'];
 
-		if(count(scandir($folder)) == 2)
+		if(is_dir($folder) && count(scandir($folder)) == 2)
 		{
 			rmdir($folder);
 			//do_log("Deleted Folder: ".$folder);
